some fixes and translate to russian

+++ app/src/main/java/com/example/weatherapp/model/common/Current.kt
@@ -5,7 +5,7 @@ data class Current(
     val feels_like : Double,
     val pressure : Int,
     val humidity : Int,
-    val uvi : Int,
+    val uvi : Double,
     val wind_speed : Double,
     val wind_deg : Int,
     val weather : List<Weather>
